new_gui/requirements.txt
zwoasi~=0.1.0.1
matplotlib~=3.6.2
setuptools~=65.5.1
numpy~=1.23.4
Pillow~=9.3.0